Create kframes_CNN_basic.py
This adds the ability to add as many kframes as you want to the basic CNN architecture that was working for Andy. 
Sad part, I cannot get it to work for 4 frames and I'm tired of banging my head. I verified that the memory is working as it should. Any help would be appreciated